requirements.txt
hikari[speedups]~=2.0.0.dev104
hikari-lightbulb==2.0.0
asyncpg
apscheduler>=3.7.0
uvloop>=0.16